feat(plotting/plotDBDict.py): add isSelectedCustomFunction hook for custom selection of analysis IDs
This commit introduces a new function `isSelectedCustomFunction` which allows for easy selection of analysis IDs based on any requirement. The function takes an analysis ID and a list of transactions as input, and returns True if the analysis ID is in a predefined set of metas (CMS and ATLAS SUSY analyses). This can be useful for selecting specific analyses for plotting or other purposes.

In addition, a new command-line argument `--use_custom_function` has been added to the script's argparser, which enables the use of this custom function for selection.